Refactor: auth package
