+++ src/main/java/iudx/data/marketplace/apiserver/provider/controller/ProviderApis.java
@@ -17,16 +17,16 @@ import io.vertx.core.json.DecodeException;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
-import iudx.data.marketplace.aaaService.AuthClient;
+import iudx.data.marketplace.dmpAuth.aaaService.AuthClient;
 import iudx.data.marketplace.apiserver.exceptions.DxRuntimeException;
 import iudx.data.marketplace.apiserver.util.RequestType;
 import iudx.data.marketplace.authenticator.service.AuthenticationService;
 import iudx.data.marketplace.authenticator.handlers.authentication.AuthHandler;
 import iudx.data.marketplace.authenticator.handlers.authentication.TokenIntrospectHandler;
 import iudx.data.marketplace.authenticator.handlers.authorization.AuthorizationHandler;
-import iudx.data.marketplace.authenticator.handlers.authorization.UserInfoFromAuthHandler;
+import iudx.data.marketplace.dmpAuth.UserInfoFromAuthHandler;
 import iudx.data.marketplace.authenticator.service.model.DxRole;
-import iudx.data.marketplace.authenticator.service.model.UserInfo;
+import iudx.data.marketplace.dmpAuth.UserInfo;
 import iudx.data.marketplace.policies.service.model.User;
 import iudx.data.marketplace.postgres.service.PostgresService;
 import iudx.data.marketplace.product.service.ProductService;
