+++ src/main/java/iudx/data/marketplace/apiserver/provider/linkedaccount/controller/LinkedAccountController.java
@@ -13,7 +13,7 @@ import io.vertx.core.json.DecodeException;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
-import iudx.data.marketplace.aaaService.AuthClient;
+import iudx.data.marketplace.dmpAuth.aaaService.AuthClient;
 import iudx.data.marketplace.apiserver.provider.linkedaccount.service.LinkedAccountService;
 import iudx.data.marketplace.apiserver.util.RequestType;
 import iudx.data.marketplace.common.ExceptionHandler;
@@ -21,9 +21,9 @@ import iudx.data.marketplace.common.ValidationHandler;
 import iudx.data.marketplace.authenticator.handlers.authentication.AuthHandler;
 import iudx.data.marketplace.authenticator.handlers.authentication.TokenIntrospectHandler;
 import iudx.data.marketplace.authenticator.handlers.authorization.AuthorizationHandler;
-import iudx.data.marketplace.authenticator.handlers.authorization.UserInfoFromAuthHandler;
+import iudx.data.marketplace.dmpAuth.UserInfoFromAuthHandler;
 import iudx.data.marketplace.authenticator.service.model.DxRole;
-import iudx.data.marketplace.authenticator.service.model.UserInfo;
+import iudx.data.marketplace.dmpAuth.UserInfo;
 import iudx.data.marketplace.authenticator.service.AuthenticationService;
 import iudx.data.marketplace.common.Api;
 import iudx.data.marketplace.common.HttpStatusCode;
