+++ src/main/java/iudx/data/marketplace/consumer/controller/ConsumerApis.java
@@ -15,15 +15,15 @@ import io.vertx.core.json.DecodeException;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
-import iudx.data.marketplace.aaaService.AuthClient;
+import iudx.data.marketplace.dmpAuth.aaaService.AuthClient;
 import iudx.data.marketplace.apiserver.util.RequestType;
 import iudx.data.marketplace.authenticator.handlers.authentication.AuthHandler;
 import iudx.data.marketplace.authenticator.handlers.authentication.TokenIntrospectHandler;
 import iudx.data.marketplace.authenticator.handlers.authorization.AuthorizationHandler;
-import iudx.data.marketplace.authenticator.handlers.authorization.UserInfoFromAuthHandler;
+import iudx.data.marketplace.dmpAuth.UserInfoFromAuthHandler;
 import iudx.data.marketplace.authenticator.service.AuthenticationService;
 import iudx.data.marketplace.authenticator.service.model.DxRole;
-import iudx.data.marketplace.authenticator.service.model.UserInfo;
+import iudx.data.marketplace.dmpAuth.UserInfo;
 import iudx.data.marketplace.common.*;
 import iudx.data.marketplace.consumer.service.ConsumerService;
 import iudx.data.marketplace.policies.service.model.User;
