+++ src/main/java/iudx/data/marketplace/dmpAuth/aaaService/AuthClient.java
@@ -1,4 +1,4 @@
-package iudx.data.marketplace.aaaService;
+package iudx.data.marketplace.dmpAuth.aaaService;
 
 import static iudx.data.marketplace.apiserver.util.Constants.*;
 import static iudx.data.marketplace.common.HttpStatusCode.INTERNAL_SERVER_ERROR;
@@ -9,7 +9,7 @@ import io.vertx.core.buffer.Buffer;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.client.HttpResponse;
 import io.vertx.ext.web.client.WebClient;
-import iudx.data.marketplace.authenticator.service.model.UserInfo;
+import iudx.data.marketplace.dmpAuth.UserInfo;
 import iudx.data.marketplace.policies.service.model.User;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
